+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Fixed:
 - Fixed player rendering on manual autocannon, including invisibility on Sable sublevels
 - Fixed shot balls not present in the Create Big Cannons creative tab
 - Fixed player rotation on Fixed Cannon Mount manual autocannon
+- Fixed #create:safe_nbt exploits
 
 ## [5.11.2] - 2026-02-27